refactor(web): eventOutputTarget moved to engine/attachment
