+++ web/src/engine/attachment/src/index.ts
@@ -1,2 +1,3 @@
 export { AttachmentInfo } from './attachmentInfo.js';
+export { eventOutputTarget, outputTargetForElement } from './outputTargetForElement.js';
 export { PageContextAttachment } from './pageContextAttachment.js';
